backend/requirements.txt
fastapi
uvicorn[standart]
sqlalchemy[asyncio]
asyncpg
PyJWT
passlib[bcrypt]
python-dotenv